+++ redditapi/src/main/java/com/worker8/redditapi/RedditApi.kt
@@ -5,16 +5,17 @@ import com.github.kittinunf.fuel.httpGet
 import com.github.kittinunf.fuel.rx.rx_object
 import com.github.kittinunf.result.Result
 import com.google.gson.GsonBuilder
+import com.worker8.redditapi.model.t3_link.RedditLinkListing
 import io.reactivex.Observable
 
 class RedditApi(val subreddit: String = defaultSelectedSubreddit) {
 
     val REDDIT_API_BASE = "https://www.reddit.com/r/"
     var after = ""
-    fun getMorePosts(): Observable<Result<Listing, FuelError>> =
+    fun getMorePosts(): Observable<Result<RedditLinkListing, FuelError>> =
         "$REDDIT_API_BASE$subreddit.json?after=$after"
             .httpGet()
-            .rx_object(Listing.Deserializer())
+            .rx_object(RedditLinkListing.Deserializer())
             .toObservable()
             .doOnNext { (listing, fuelError) ->
                 listing?.value?.after?.also { after = it }
