--- redditapi/src/main/java/com/worker8/redditapi/RedditLink.kt
@@ -1,9 +1,0 @@
-package com.worker8.redditapi
-
-import com.google.gson.annotations.SerializedName
-
-data class RedditLink(
-        @SerializedName("data")
-        val value: RedditLinkData = RedditLinkData(),
-        val kind: String = ""
-)
