+++ redditapi/src/main/java/com/worker8/redditapi/model/t3_link/RedditLink.kt
@@ -1,0 +1,11 @@
+package com.worker8.redditapi.model.t3_link
+
+import com.google.gson.annotations.SerializedName
+
+data class RedditLink(
+    @SerializedName("data")
+    val value: RedditLinkData = RedditLinkData(),
+
+    @SerializedName("kind")
+    val kind: String = "" // kind = t3, Link
+)
